[MOVEONLY] Move sigencoding checker functions
Similar to https://reviews.bitcoinabc.org/D1565 by deadalnix

This eases fetching new changes from ABC.
