+++ src/script/script_error.h
@@ -78,4 +78,18 @@ typedef enum ScriptError_t
 
 const char* ScriptErrorString(const ScriptError error);
 
+inline bool set_success(ScriptError* ret)
+{
+    if (ret)
+        *ret = SCRIPT_ERR_OK;
+    return true;
+}
+
+inline bool set_error(ScriptError* ret, const ScriptError serror)
+{
+    if (ret)
+        *ret = serror;
+    return false;
+}
+
 #endif // BITCOIN_SCRIPT_SCRIPT_ERROR_H
